Myriad changes around running workers.
First, added support for detecting when a module has been updated, and then reloading
that module in the worker. This makes the development cycle a lot easier.

Second, added support for filesystem sandboxing. Every worker is run in its own working
directory sandbox, and wipes it clean before running a job, and after, to ensure that
no cross-job contamination takes place.