Initial support for resumable jobs, Also, bug fix.
It turns out that we weren't creating a new client for each spawned worker, and
it's possible to get race conditions when sharing a file descriptor between
processes. I was seeing the result of 'pop' requests as the results for load
scripts, and vice-versa. It's also important to instantiate the queues /after/
the new client is instantiated.